Added channel closure after sending message
