--- atgs_app/android/app/src/main/java/com/example/atgs_app/RemoteGlanceWidgetReceiver.kt
@@ -1,7 +1,0 @@
-package com.example.atgs_app
-
-import HomeWidgetGlanceWidgetReceiver
-
-class RemoteGlanceWidgetReceiver : HomeWidgetGlanceWidgetReceiver<RemoteGlanceWidget>() {
-    override val glanceAppWidget = RemoteGlanceWidget()
-}
