--- atgs_app/android/app/src/main/res/xml/file_paths.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<paths xmlns:android="http://schemas.android.com/apk/res/android">
-    <files-path
-        name="internal_files"
-        path="." />
-</paths>
