--- atgs_app/android/app/src/main/res/xml/remote_glance_widget.xml
@@ -1,6 +1,0 @@
-<appwidget-provider xmlns:android="http://schemas.android.com/apk/res/android"
-    android:initialLayout="@layout/glance_default_loading_layout"
-    android:minWidth="203dp"
-    android:minHeight="90dp"
-    android:resizeMode="horizontal|vertical"
-    android:updatePeriodMillis="10000" />
