+++ atgs_app/android/app/src/main/res/xml/remote_widget_info.xml
@@ -3,11 +3,12 @@
     android:description="@string/app_widget_description"
     android:initialKeyguardLayout="@layout/remote_widget"
     android:initialLayout="@layout/remote_widget"
-    android:minWidth="180dp"
-    android:minHeight="180dp"
+    android:minWidth="110dp"
+    android:minHeight="40dp"
+    android:previewImage="@drawable/example_appwidget_preview"
     android:previewLayout="@layout/remote_widget"
     android:resizeMode="horizontal|vertical"
-    android:targetCellWidth="3"
-    android:targetCellHeight="3"
+    android:targetCellWidth="2"
+    android:targetCellHeight="1"
     android:updatePeriodMillis="86400000"
     android:widgetCategory="home_screen" />
